Refactor and relocate document upload components
Moved document upload components from 'Belgeler' to 'Dosya' directory and updated all relevant imports. Unified and refactored file upload logic, removed unused delete functionality, and improved consistency in file and image upload APIs. Also updated 'EditDrawer' to standardize KDV field values and cleaned up image upload code.
